add coco-caption evaluation code. add visual genome 1.2 processing scripts. refactor model prototxt for compactness. make initial weights of networks to be Gaussian (variance 0.01). refactor caption evaluation code to make it compatible with different model structures.

+++ lib/rpn/proposal_target_single_class_layer.py
@@ -56,7 +56,7 @@ class ProposalTargetLayer(caffe.Layer):
 
         num_images = 1
         rois_per_image = cfg.TRAIN.BATCH_SIZE / num_images
-        fg_rois_per_image = np.round(cfg.TRAIN.FG_FRACTION * rois_per_image)
+        fg_rois_per_image = int(np.round(cfg.TRAIN.FG_FRACTION * rois_per_image))
 
         # Sample rois with classification labels and bounding box regression
         # targets
